+++ test-framework/core/Codecs/ramonly/RAMOnlyPostingsFormat.cs
@@ -649,7 +649,7 @@ namespace Lucene.Net.Codecs.ramonly
 	  // Holds all indexes created, keyed by the ID assigned in fieldsConsumer
 	  private readonly IDictionary<int?, RAMPostings> State = new Dictionary<int?, RAMPostings>();
 
-	  private readonly AtomicInteger NextID = new AtomicInteger();
+	  private readonly AtomicLong NextID = new AtomicLong();
 
 	  private readonly string RAM_ONLY_NAME = "RAMOnly";
 	  private const int VERSION_START = 0;
